Attempting to deal with bulk issues, these are disconnects in the patcher and midi timing issues.
Attempted the following:

using the fifo pump thread has issues under high load.

1. With HIGHPRIO it affects the DSP, crackles on patches as process() is interrupted.
2. With HIGHPRIO-1 it can get starved, dead Ksoloti as bulk is not working restart required.

The fifo pump cannot be directly run in the SAI interrupt as it doesn’t work correctly, not sure why.
The fifo pump cannot be run in the patch thread as this thread can block.

We need a two pronged approach

When patch is running use external fifo pump.
When patch is not running use existing thread.

This doesn’t help. I’m beginning to think that the bulk and midi drivers need rewriting to work correctly with V16.
The queues need totally removing.